fixed ai comments
