+++ src/app/portfolio/components/PortfolioHistory.tsx
@@ -13,7 +13,6 @@ import {
 import debounce from 'lodash.debounce'
 import React, { memo, PropsWithChildren, useCallback } from 'react'
 import InfiniteScroll from 'react-infinite-scroll-component'
-import { v4 as uuidv4 } from 'uuid'
 import Loader from '@/components/common/loader'
 import Skeleton from '@/components/common/skeleton'
 import { PortfolioHistoryRedeemItem } from '@/app/portfolio/components/PortfolioHistoryRedeemItem'
@@ -46,7 +45,7 @@ const TableContainerWrapper = ({ children }: PropsWithChildren) => {
   )
 }
 
-const History = ({ userMenuLoading }: { userMenuLoading: boolean }) => {
+const History = () => {
   const {
     data: historyData,
     fetchNextPage,
@@ -101,9 +100,15 @@ const History = ({ userMenuLoading }: { userMenuLoading: boolean }) => {
       <TableContainerWrapper>
         {historyFlat.map((item) =>
           'strategy' in item ? (
-            <PortfolioHistoryTradeItem key={uuidv4()} trade={item as HistoryTrade} />
+            <PortfolioHistoryTradeItem
+              key={(item as HistoryTrade).transactionHash}
+              trade={item as HistoryTrade}
+            />
           ) : (
-            <PortfolioHistoryRedeemItem key={uuidv4()} redeem={item as HistoryRedeem} />
+            <PortfolioHistoryRedeemItem
+              key={(item as HistoryRedeem).transactionHash}
+              redeem={item as HistoryRedeem}
+            />
           )
         )}
       </TableContainerWrapper>
